load js code via sub pages element once
